chore(Participants): move files in a single directory
