+++ src/components/RightSidebar/Participants/ParticipantsList.vue
@@ -32,8 +32,8 @@
 
 <script>
 
-import Participant from './Participant/Participant.vue'
-import LoadingPlaceholder from '../../../LoadingPlaceholder.vue'
+import Participant from './Participant.vue'
+import LoadingPlaceholder from '../../LoadingPlaceholder.vue'
 
 export default {
 	name: 'ParticipantsList',
